+++ Changelog.md
@@ -2,16 +2,14 @@
 
 Language Features:
 
-
 Compiler Features:
  * Yul: Raise warning for switch statements that only have a default and no other cases.
  * Output compilation error codes.
 
-
 Bugfixes:
  * SMTChecker: Fix internal error when encoding tuples of tuples.
  * SMTChecker: Fix aliasing soundness after pushing to an array pointer.
-
+ * Type system: Fix internal compiler error on calling externally a function that returns variables with calldata location.
 
 ### 0.6.9 (2020-06-04)